feat(graph): move DataGraph interface, remove chromaticScale from store
