+++ src/graph/graph-store.ts
@@ -1,12 +1,12 @@
 import create from 'zustand'
-import { DataGraph, getBaseGraph, getMergedGraph } from './Graph'
+import { getBaseGraph, getMergedGraph } from './Graph'
 import {
   highlightHighlight,
   collectParentColumnAndTableIds,
   collectChildColumnAndTableIds,
   notEmpty,
 } from './graph-ops'
-import { ChromaticScale, Graph, Badge, Highlight, Table } from './model'
+import { ChromaticScale, Graph, DataGraph, Badge, Highlight, Table } from './model'
 
 const defaultHighlightedColumns = {
   selectedTableId: '',
@@ -96,14 +96,11 @@ export const useGraphStore = create<GraphStore>((set, get) => ({
   closeDrawer: () => set({ drawerNodeId: undefined }),
   openDrawer: (drawerNodeId: string) => set({ drawerNodeId }),
 
-  chromaticScale: 'interpolatePuOr',
-  setInitialStore: ({ graph, data, tables, chromaticScale }) =>
-    set({ graph, data, tables, chromaticScale }),
+  setInitialStore: ({ graph, data, tables }) => set({ graph, data, tables }),
 
   selectedNodeId: undefined,
-  nodeClick: (id: string) => {
+  nodeClick: (id: string, chromaticScale: ChromaticScale) => {
     const graph = get().graph
-    const chromaticScale = get().chromaticScale
     const highlightMode = get().highlightMode
     const newHighlights = highlightHighlight(highlightMode)(graph, id, chromaticScale)
     return set({
@@ -143,7 +140,7 @@ export const useGraphStore = create<GraphStore>((set, get) => ({
   setBadge: (badge: Badge) => set({ badge }),
 
   highlightMode: 'nearest',
-  setHighlightMode: (highlightMode: Highlight) => {
+  setHighlightMode: (highlightMode: Highlight, chromaticScale: ChromaticScale) => {
     if (highlightMode === 'diffing') {
       // TODO: This block possibly does not belong here
       const baseRelease = get().baseRelease
@@ -168,7 +165,6 @@ export const useGraphStore = create<GraphStore>((set, get) => ({
     }
 
     const graph = get().graph
-    const chromaticScale = get().chromaticScale
     const newHighlights = highlightHighlight(highlightMode)(graph, selectedNodeId, chromaticScale)
     return set({
       highlightMode,
@@ -229,11 +225,10 @@ export interface GraphStore {
   closeDrawer: () => void
   openDrawer: (id: string) => void
 
-  chromaticScale: ChromaticScale
   setInitialStore: (arg: InitialArgs) => void
 
   selectedNodeId?: string
-  nodeClick: (id: string) => void
+  nodeClick: (id: string, chromaticScale: ChromaticScale) => void
   searchNodeLabels: (value: string) => string[]
 
   baseRelease: string
@@ -247,7 +242,7 @@ export interface GraphStore {
   resetHighlights: () => void
 
   highlightMode: Highlight
-  setHighlightMode: (_: Highlight) => void
+  setHighlightMode: (value: Highlight, chromaticScale: ChromaticScale) => void
 }
 
 interface HColumns {
@@ -262,7 +257,6 @@ interface InitialArgs {
   graph: Graph
   data: DataGraph[]
   tables?: Table[]
-  chromaticScale: ChromaticScale
 }
 
 interface StoredHighlights {
